test: move fixture to avoid spurious build
Move the fixture so that it is only evaluated if we are running
functional tests.
